refactor: 컴포넌트명  CustomErrorBoundary -> ErrorBoundary 수정
