+++ src/pages/DustMap.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@chakra-ui/react';
 import { ErrorFallback } from '@/components/common';
-import CustomErrorBoundary from '@/components/common/CustomErrorBoundary';
+import ErrorBoundary from '@/components/common/ErrorBoundary';
 import Map from '@/components/Map';
 import { fallbackSize } from '@/types/error';
 import { ERROR_MESSAGE } from '@/utils/constants/message';
@@ -8,7 +8,7 @@ import { ERROR_MESSAGE } from '@/utils/constants/message';
 const DustMapPage = () => {
   return (
     <Box maxWidth="30rem" height="100vh" margin="0 auto">
-      <CustomErrorBoundary
+      <ErrorBoundary
         rejectFallback={
           <ErrorFallback
             size={fallbackSize.FULL}
@@ -17,7 +17,7 @@ const DustMapPage = () => {
         }
       >
         <Map />
-      </CustomErrorBoundary>
+      </ErrorBoundary>
     </Box>
   );
 };
